// src/generated/types.ts
export type Scalars = {
	ID: string
	String: string
	Boolean: boolean
	Int: number
	Float: number
	/** Blockchain data scalar type */
	BlockchainData: any
	/** Broadcast scalar id type */
	BroadcastId: any
	/** ChainId custom scalar type */
	ChainId: any
	/** collect module data scalar type */
	CollectModuleData: any
	/** Contract address custom scalar type */
	ContractAddress: any
	/** create handle custom scalar type */
	CreateHandle: any
	/** Cursor custom scalar type */
	Cursor: any
	/** The javascript `Date` as string. Type represents date and time as the ISO Date string. */
	DateTime: any
	/** Ens custom scalar type */
	Ens: any
	/** Ethereum address custom scalar type */
	EthereumAddress: any
	/** follow module data scalar type */
	FollowModuleData: any
	/** handle custom scalar type */
	Handle: any
	/** handle claim id custom scalar type */
	HandleClaimIdScalar: any
	/** Internal publication id custom scalar type */
	InternalPublicationId: any
	/** jwt custom scalar type */
	Jwt: any
	/** limit custom scalar type */
	LimitScalar: any
	/** Locale scalar type */
	Locale: any
	/** Markdown scalar type */
	Markdown: any
	/** mimetype custom scalar type */
	MimeType: any
	/** Nft ownership id type */
	NftOwnershipId: any
	/** Nonce custom scalar type */
	Nonce: any
	/** The notification id */
	NotificationId: any
	/** ProfileId custom scalar type */
	ProfileId: any
	/** Publication id custom scalar type */
	PublicationId: any
	/** The publication tag */
	PublicationTag: any
	/** Publication url scalar type */
	PublicationUrl: any
	/** reference module data scalar type */
	ReferenceModuleData: any
	/** Query search */
	Search: any
	/** Relayer signature */
	Signature: any
	/** Sources custom scalar type */
	Sources: any
	/** timestamp date custom scalar type */
	TimestampScalar: any
	/** The tx hash */
	TxHash: any
	/** The tx id */
	TxId: any
	/** UnixTimestamp custom scalar type */
	UnixTimestamp: any
	/** Url scalar type */
	Url: any
	/** Represents NULL values */
	Void: any
}
// ...
export type CreateAskBroadcastItemResult = {
	__typename?: 'CreateAskBroadcastItemResult'
	/** The date the broadcast item expiries */
	expiresAt: Scalars['DateTime']
	/** This broadcast item ID */
	id: Scalars['BroadcastId']
	/** The typed data */
	typedData: AskPostEip712TypedData
}

/** The create post eip 712 typed data */
export type AskPostEip712TypedData = {
	__typename?: 'CreatePostEIP712TypedData'
	/** The typed data domain */
	domain: Eip712TypedDataDomain
	/** The types */
	types: CreatePostEip712TypedDataTypes
	/** The values */
	value: CreatePostEip712TypedDataValue
}

/** The create post eip 712 typed data types */
export type CreatePostEip712TypedDataTypes = {
	__typename?: 'CreatePostEIP712TypedDataTypes'
	PostWithSig: Array<Eip712TypedDataField>
}

/** The create post eip 712 typed data value */
export type CreatePostEip712TypedDataValue = {
	__typename?: 'CreatePostEIP712TypedDataValue'
	collectModule: Scalars['ContractAddress']
	collectModuleInitData: Scalars['CollectModuleData']
	contentURI: Scalars['PublicationUrl']
	deadline: Scalars['UnixTimestamp']
	nonce: Scalars['Nonce']
	profileId: Scalars['ProfileId']
	referenceModule: Scalars['ContractAddress']
	referenceModuleInitData: Scalars['ReferenceModuleData']
}
// ...
export type Eip712TypedDataDomain = {
	__typename?: 'EIP712TypedDataDomain'
	/** The chainId */
	chainId: Scalars['ChainId']
	/** The name of the typed data domain */
	name: Scalars['String']
	/** The verifying contract */
	verifyingContract: Scalars['ContractAddress']
	/** The version */
	version: Scalars['String']
}
// ...
export type RelayError = {
	__typename?: 'RelayError'
	reason: RelayErrorReasons
}

/** Relay error reason */
export enum RelayErrorReasons {
	Expired = 'EXPIRED',
	HandleTaken = 'HANDLE_TAKEN',
	NotAllowed = 'NOT_ALLOWED',
	Rejected = 'REJECTED',
	WrongWalletSigned = 'WRONG_WALLET_SIGNED',
}

export type RelayResult = RelayError | RelayerResult

/** The relayer result */
export type RelayerResult = {
	__typename?: 'RelayerResult'
	/** The tx hash - you should use the `txId` as your identifier as gas prices can be upgraded meaning txHash will change */
	txHash: Scalars['TxHash']
	/** The tx id */
	txId: Scalars['TxId']
}